Remove separate setup.py file for Python 3.
